Reservations functionality (70%)
- Added reservation screens
- Finished cart spinners

+++ PPI_MOBILE/viroshop/lib/World/CartItem.dart
@@ -4,6 +4,6 @@ import 'package:viroshop/World/Product.dart';
 class CartItem {
   final int id;
   final Product cartProduct;
-  final double quantity;
+  final int quantity;
   CartItem(this.id, this.cartProduct, this.quantity);
 }
